requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
twilio==9.8.1
sendgrid==6.11.0
schedule==1.2.0
httpx==0.26.0
supabase==2.8.1
postgrest==0.17.1
python-dotenv==1.0.0
jinja2==3.1.2
email-validator==2.1.0
bcrypt==4.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# WhatsApp Chat Processing
spacy==3.7.2
python-dateutil==2.8.2
dateparser==1.2.0
regex==2023.10.3